Finish security implementations

+++ src/main/resources/db/migration/V7__Create_Table_State.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS `state` (
-   `id` smallint unsigned NOT NULL AUTO_INCREMENT,
+   `id` BIGINT(20) NOT NULL AUTO_INCREMENT,
    `name` varchar(255) DEFAULT NULL,
-   `country_id` smallint unsigned NOT NULL,
+   `country_id` BIGINT(20) NOT NULL,
    PRIMARY KEY (`id`)
  );
